Update biomass and residue cover based on crop
Lines 287-313.  Not sure if I need to pull the source of the crop type... I believe that the way the crop type is pulled from line 30 that I can call it as it is written (starting in line 287), but I am not sure if this is true.  This is currently only written for corn and soybean, will need to find representative values for the "other crops" category.